requirements.txt
# all codes are run with Python 3.9.15 and CUDA 11.6
# ML
torch==1.13.1
torchvision==0.14.1
torch-scatter==2.1.0
torch-geometric==2.2.0

# scientific
numpy==1.23.4
scipy==1.9.3
Pillow==9.3.0
matplotlib==3.6.2

# utils
tqdm==4.64.1
PyYAML==6.0
easydict==1.10
tensorboard==2.11.0

# inference, only needed for running the Jupyter notebooks
jupyterlab==3.5.2
